feat: code structure adjustment
