+++ packages/record/src/watchers/index.ts
@@ -1,0 +1,6 @@
+export * from './dom-watcher'
+export * from './form-element-watcher'
+export * from './location-watcher'
+export * from './mouse-watcher'
+export * from './scroll-watcher'
+export * from './window-watcher'
