+++ packages/utils/src/index.ts
@@ -10,3 +10,4 @@ export * from './scroll'
 export * from './output'
 export * from './transform'
 export * from './performance/fmp'
+export * from './watcher-tool'
